Create Stock::SimpleCoordinator
This class is intended to replace Spree::Stock::Coordinator as well as its helper classes:
  Stock::Adjuster
  Stock::Packer
  Stock::Prioritizer

This solves several issues and bugs with the existing Coordinator, where
stock coordination would fail despite actually having enough stock. It
should also be faster than the existing Coordinator by making use of the
new Stock::Availability class.

My hope is that this class is simple enough that it could be copied in
to stores and used as a starting point to customizations.