Use ducktyping in wallet payment source validation
Instead of asking for the source class to be a `Spree::PaymentSource`
we should only ask if it responds to `reusable?`; as this method is actually
the only method we expect to be implemented on the source.

Otherwise payment sources that do not inherit from `Spree::PaymentSource`
(a class we first introduced with Solidus 2.2) cannot be used with the wallet.