Fix spec dependant on database ordering
This spec previously created two line_items, and relied on the order
they came back from the DB (and only testing the first). I believe the
second line item was not intended by the original author of this spec.

This commit removes the second line item and makes the spec consistent
regardless of DB ordering.