Speed up taxation of shipping rates
Instead of loading the tax rates for the order for each rate which needs
taxing, we can perform the SQL queries only once.

This is done by building a single instance of the
shipping_rate_tax_calculator_class for a shipment, and calling
`calculate` on that object for each rate, allowing it to cache the order
rates.