Do a solidus 1.4 roll-up migration
The logical break between solidus 1.4 and 2.0 means we can take some
liberties with some clean up in a pretty safe way.

This rolls up the migrations (something we haven't done for the last 5
years), cleaning up our migration directory and importing fewer
migrations into new stores.

It also has the added benefit of not needing to run through the entire
migration history for a test_app, or having to maintain any of the old
migrations.

Like the spree_one_two rollup, I had to include a definition of
spree_users even though it shouldn't be part of core's schema.