Do not validate the format of zipcodes
We shouldn't validate the format of zipcodes. They are vastly different
across countries, and the current validation does not tell us anything
about whether the address can be shipped to, it only checks the format
of the zipcode entered using a regular expresssion.

I've briefly thought about adding a customizable address validator class,
but that would've prompted users to hook address validation that hits the
internet into the ActiveRecord callback chain. We can not want that.

The replacement for this must be to validate addresses from host apps.
Preferably this takes place at the interaction level rather than at the
persistence level, as it is done here.

The format validation using Twitter CLDR comes at a cost of a 60MB gem.