Use form submit events instead of click events
In order to give users the same UX they know from form fields, we need to wrap the tracking input field into a form and use a submit button as save button.

Because the submit buttons sits outside of the form and we can't wrap the whole table row in a form (HTML semantics) we need to use the barely known form attribute of submit buttons.

Now we can use the form submit event instead of the save button click event and one can submit the form by clicking the save button or by hitting the enter key inside the form.

Use form submit event for saving order shipment