Styling fixes for number with currency component
The amount needs a min-width in order to not be too small in narrow areas. One might not be able to see the value.

The addon shoud not have a fixed width, but a min-width so it grows when necessary, but not take too much space it actually does not need.