Re-arrange order payments table
The order payments table has lots of information and lacks some structure.

By moving the amount to the right of the table and give it some as well as align it to the right this important piece of information is now better visible. If a user wants to change the amount of an uncaptured payment the form is now right by the action icons and not somewhere in the middle.

By further shorten the creation date format we gain some valuable extra space.