Add required to product master price field
Depending on Spree::Config.require_master_price, this adds
`required: true` to the master price field on the products form, as well
as the indicator to the field label. When config is set to false, the
field can be cleared, and the form submitted - resulting in the price
being defaulted to `0.00`

This is part of issue #1861. Though in the issue report it talks of the
product price showing a deleted value when removed (which has been
resolved). This further makes it a conscious decision to set the price
to `0`.