Do not use z-index for aligning strike through headlines
Using `z-index` to align objects on the z axis is problematic
and hard to get right across the whole admin.
We should rely on the DOM position instead. In case of the `strike-through`
headline mixin we do not need to change anything because the pseudo element
already sits below the headline.