Support .environment on Spree::Config
Currently, we have two major way of configuring solidus.

The first is the reasonably-clear Spree::Config, which is an instance of
Spree::AppConfiguration and contains a number of configuration settings.

Secondly is the Spree::Core::Environment, which until this commit was
only stored and could only be accessed through
Rails.application.config.spree, though there was no clear distinction on
why some settings were here and others in Spree::Config.

This takes a step forward to clarify those settings, allowing access to
the Spree::Core::Environment settings through
`Spree::Config.environment`