Pull environment setup out of rails initializers
Setting these "default" values during rails initializers is confusing,
as understanding and/or debugging rails initializers is difficult and
time consuming.

By pushing the setup into the Spree::Config, we can simplify the setup
of the environment and remove the confusion around when and in what
order these fire.

The old initializers are left so existing applications and extensions
can hook into them, though they should not notice a difference.