Deprecate class method in `Calculator::FlexiRate`
This class method is not used anywhere in the code base. As a
calculator, the `available?` method should in general be an instance
method (which defaults to `true` from the inherited base calculator class).

I can't go back far enough in the history of the source code to find out
why this method was added in the first place.