Avoid extra query on ShppingRate#display_price
We were checking taxes.empty? and if it wasn't we were mapping over the
list of taxes. Instead, we can call taxes.to_a.empty?, if there are no
taxes this is just an empty array and if there are we load them and
tests if there are any with just the one query.