Bump ActiveMerchant Dependency to 1.66 minimum
Versions before that do not work with Rails 5, upon which we depend.

With recent work on explicitly requiring only those bits of ActiveMerchant
that we need, I needed to test the full range of ActiveMerchant versions
for compatibility in their gem structure.

That task can only be started at the first version our gem actually bundles with.