Don't modify DB after render in order_details_spec
Previously this spec would make modifications to the DB either after or
during the render of the page it was testing.

This happened to work under our current implementation, but has
potential to cause issues as our splitting interface becomes smarter.

It also sometimes caused SQLite3::BusyException if testing under sqlite.