Use transactional fixtures in frontend and backend
Previously, we needed to run tests outside of a transaction because our
test and server thread required separate database connections.

However as of Rails 5.1, due to the work implementing system tests, this
is no longer necessary.

This should be significantly faster for postgresql and mysql, which
otherwise had to truncate all tables after every feature spec.