Add `#gallery` to `Variant` and `Product`
`Product` and `Variant` should not know the implementation details of
their images. When they have that knowledge we end up coupling the image
uploading library (like Paperclip) way too tightly to these core models.

Developers are able to write their own custom gallery classes that
respond to a single `#images` method that returns a collection of images
that implement the interface defined in `Spree::Image`.