Remove testing_support/i18n
This is a file which was used to detect missing or unused translations,
but it is no longer useful.

It only detects translations made using Spree.t, which we no longer use,
and haven't used exclusively in years. I don't believe the detection of
unused translations has ever worked, and the detection of missing
translations can use raise_on_missing_translations instead.