Add image galleries for `Product` and `Variant`
Paperclip is deprecated, we need to remove it from being stock Solidus.
This becomes a problem because we have very specific Paperclip method
calls in our view layer, and generally are too tightly coupled to
everything Paperclip related.

Having a pluggable gallery implementation for products and variants
gives us better control of how we send things to the view layer and can
hide implementation details of specific image uploading libraries. If a
developer wants to plug a different image library they need to write a
custom gallery implementation that returns a collection of images that
respond to the same interface as `Spree::Image`.

See PR 2337 for more context.