Admin order edit: Clear address state_id when switching to country w/o states
The admin wasn't clearing `address.state_id` from the DB when you switched from
a country with states to a country without states.  You would then end up with
an address with both `state_name` and `state_id`, and `state_text` would end up
showing the old `state_id`.

This happened because we have the complicated state_name + state_id situation
with addresses, and we swap out one form input with another depending on the
country.

We needed to provide default hidden inputs for state_name and state_id with
blank values.