Move role configuration fully into Spree::Config
Previously, RoleConfiguration was a singleton class, which was
configured by engine initialization.

Instead, we can configure it like the rest of our configuration by
moving it to AppConfiguration (and therefore Spree:Config). It is also
no longer a Singleton class, which makes testing it simpler.

This deprecates RoleConfiguration.instance and
RoleConfiguration.configure and has them use Spree::Config.roles
instead.