Fix hovering issues on tables actions links
Clicking on an action button in a table, a new template for the row
is rendered under current mouse coordinates. This was not making fire
the mouseleave callback for that element.

This commit introduces a MutationObserver that takes the same mouseleave
actions when elements inside that row changes. Also, move the mouseleave
event listener into the mouseenter callback. This way we have the
instance of the MutationObserver and we can disconnect it properly to
avoid aving several mutationobserver attached to the same elements due
to several consecutive mouseenter/mouseleave on that row.