Recalculate order inside UnitCancel adjust!
Order is being recalculated every time a new adjustment is added to the
line_item. There could be miscalculations when multiple inventory_units
were canceled in a loop. Now the order is recalculated and the
line_item.total is representative of the current state. Add new spec
for UnitCancel adjust!