Adjust dummy_app initialization
Previously we initialized the dummy_app implicitly when requiring the
file. Because of this we were relying on some hacks like using
ENV['BUNDLE_GEMFILE'] to try and determine the path to the desired
spec/dummy.

To be more reliable and to allow a single Gemfile, this commit requires
the app to be explcitly initialized through a new DummyApp.setup method.