Use new image interface in backend image partial
Instead of calling a Paperclip specific method (`#attachment`) we
instead want to use the more generic interface that we have now defined
in `Spree::Image`. This way as long as our variant and product galleries
return images that respond to a common interface we can swap out our
image uploading libraries as we wish.