Remove span.required in favor of adding * with css
Remove that extra span element in favor of using css to add this
sign near the required fields label. This needed to add a class
(field-required) to required fields contanier div.

JS address is also changed so that instead of toggling span.required
visibility it toggle field-required class on field container.