Remove bottom border of non-form fieldsets
We have too many borders in our layout. Whitespace, tables and headers are separating content areas already, no need additional borders. That only distracts the eye.

Keeps the bottom border for fieldsets that are holding a form. This makes sense as the form buttons sit on top of the bottom border and visually encloses the form field set.