Rewrite getting started article
I rewrote the getting started guide to focus on first-time
installations. Other "getting started"-related information can reside
in this subdirectory alongside this article.

The article is opinionated in that it focuses on macOS users. The
rationale for this is that a) Linux users are more likely to know how
to install the required system dependencies and don't need that
hand-holding and b) Windows users have to jump through additional hoops
to get Solidus up and running (we don't recommend this).

That said, there are still links and resources for Linux users
available in this article.