Allow individual suites to be run from top-level
This allows running individual projects' spec suites from the top-level.
For example:

    rake spec:core

Will run the spec suite for the core project.

This also defines tasks like db:reset:core, but doesn't give them a
description to avoid cluttering up rake -T.

I don't think we need to add this to the README, as the existing method
still works and is probably more flexible. This is just another
convenient way to do the same.