Avoid changing method visibility when deprecating a method
ActiveSupport::Deprecation::MethodWrapper#deprecate_methods changes method
visibility to `public`.  This seems unintentional and I will submit a PR to
Rails to see if that can be updated.

For the meanwhile, this update changes method visibility back to the original
visibility. It's a bit more convoluted than might be expected because
ActiveSupport uses `Module.prepend` to override the original method, so we can't
just write `private :some_method` because that only affects the original method,
whose visibility has not been changed.