Add `#url` and `#filename` to Spree::Image
The view layer in Solidus uses too many many Paperclip specific calls.
In order to remove Paperclip from core (since it is no longer being
maintained) we need to move those calls out of the view layer and into
here. The idea being that a different image provider need only to give
the view layer images that respond to the same interface defined here.

The image galleries (like Spree::Gallery::ProductGallery) should respond
to `#images` with a collection of objects that respond to `#id`,
`#filename`, `#url(size)`, and `#viewable_id`.

Instead of removing Spree::Image and creating a new class to represent
this image class for the view layer, we can just transition Spree::Image
into that role.