Use scope.cache_key instead of doing it ad-hoc
This was written before rails supported calling cache_key on a scope.
This saves us one query, since rails selects both the max and count in a
single query.

This removes some assertions which tested explicitly for the cache key
which was written. This should be covered by the assertions in the same
specs which test for how many cache writes are performed.