Remove broken deprecation on Spree::PaymentMethod#provider_class
Simply calling `deprecate` created the `provider_class` method on the base class
(and all subclasses) but it did so in a way that would blow up with an error
like:

    NoMethodError: super: no superclass method `provider_class'

Defining the method also messed up the check on `respond_to? :provider_class`
in the `gateway_class` method.

Also, calling `deprecate` in the base class would not generate deprecation
warnings for any `provider_class` methods defined in subclasses.