Document promotion flow for admins and customers
This section gives a high-level overview of what happens when a
promotion is being added in the backend, then how it gets applied when
a customer is checking out.

This example uses a free shipping promotion, since that would be a
very typical use case for promotions. There is still room for examples
that create other types of discounts.