Hedge bets in docs on overriding assets
At some point in the future it's possible we will use an asset bundler
other than sprockets (like webpack or rollup). If we do so we won't be
able to support sprocket's view-like file lookup and overriding (and it
is an unusual thing to support for JS in the first place).