Replace new product inline form with normal form
Previously, creating a new product involved filling out a small inline
form which held with only the required attributes of a product.

This requires admin users to understand and learn to use two separate
forms with the same purpose (filling in product data). It's also awkward
if the user wants to fill in some of the information not on this page
as they have to create the product.

This commit removes the inline new product form in favour of linking to
a new product page with the same form as the edit page.