Use #instance_of? to check calculator class
Using #is_a? to check the calculator class doesn't play well with calculators that are subclasses of each other. For example, given calculator A and calculator B, where B is a subclass of A: if a shipping method is configured to use calculator B, it becomes impossible to switch the shipping method to calculator A since A.is_a?(B) returns true. Whereas A.instance_of?(B) returns false. If the intention is to prevent the same calculator from being set twice, #instance_of? is the correct choice since it disregards subclasses and only returns true if both are the exact same class.