Localize flatpickr
This avoids needing to load extra js files (as we do for select2, and
regularly causes headaches). Instead we just use the translation keys we
already have available. This doesn't fill in all the information
flatpickr can store for a locale (ex. long day names and short month
names), but it fills in all the data we use.