Remove eager touching
Before:
Any time an option value is saved with changes (or touched) all
associated variants are also touched.

In the core test suite this is called 803 times creating roughly 2 db
requests per call.

After:
Touching variants after an option value is updated now occurs in the
admin controllers when option types or option values are modified there.

The touching mechanism remains in the model for other extensions to hook
into if they need to trigger a cache invalidation from something like an
import