Use datalist instead of jquery-ui/autocomplete
Replaces jquery-ui/autocomplete, which is only used in the product
properties page, with the html5 <datalist> element.

Browser support for datalist is patchy (notably no safari), but it is
entirely fine for this page not to have autocompletion.

We had apparently already decided that the number of properties is low
enough that they can be just dumped on the page as json, so we might as
well dump them on the page as HTML.