Use SimpleCov 'rails' preset
The "rails" helper is built into SimpleCov. It adds all the groups we
previously had (except for views, which can't be measured using
simplecov anyways). It also declares track_files on app/ and lib/, so it
will show files which are never even loaded, which previously there was
no indication of.