Use backbone StateSelect view for stock locations
This was using the old update_state function which was no longer used by
the customer address page. It worked okay, but was pretty clunky.

This commit changes the stock locations page to use the same
Views.StateSelect backbone view that the customer details page uses.
This allows removing the inline JS on that page, and removes the last
use of the update_state global.

This also makes the behaviour of Views.StateSelect better when no
country is selected. It now shows the state selector in a disabled
state.