Simplify store_credit memo edit JS
The previous store credit JS did a lot more than it needed to, and
recreated much of the work that we could get for free from rails-ujs. We
were already specifying remote: true, but weren't using that
functionality and instead were performing the AJAX explicitly.

This commit replaces the existing JS and relies on CSS for display , and
jquery-ujs for submitting via AJAX. The user-facing behaviour should be
identical.

The resulting code has very little to do with store credits, we should
consider at a future date looking through our inline edits in the admin
and possibly consolidating them.