Simplify tabs component
Previously we had logic to see whether the container had grown or
shrunk, and responded by either taking elements out of the dropdown or
moving them in.

Instead, this commit combines the separate code paths. Every time the
width is changed all elements are removed from the DOM and one-by-one
re-inserted into the correct parent.

This shouldn't be significantly slower then the previous behaviour.
We are looking at very few DOM nodes and are careful to not force layout
at any point in the loop.