Add spec demonstration actual behaviour
Since line_item level promotion adjustments are only applied to
elligible line_items, there's a surprising interaction between
the distributed amount calculator and the product promotion rule.

The calculator distributes the preffered amount among all of the
order's line_items but only discounts the order by the portion
which was applied to the line_item with the elligible product.