Rewrite overview of Solidus payments system
This commit reduces the barrier to entry for developers getting started
with Solidus's payments system.

It does this by identifying five essential parts of the payments system
and explaining how they fit together to make payments work.

I realize that there are other parts of the system that are not covered
yet. I can imagine additional articles about payments being linked to
from this overview in future pull requests.