Set model_class on admin promotion rules controller
If a user has the `PromotionManagement` permission set, they are
currently unable to add new promotion rules to promotions.

This is happening because the promotion rules controller doesn't inherit
from the `ResourceController`, so if it doesn't respond to
`:model_class`, it will try to authorize using `controller_name.to_sym`
instead of the correct class constant.