Accept switch_to_locale param in LocaleController
Previously we accepted only the :locale param in LocaleController#set.

solidus_i18n, however, only accepts the :switch_to_locale param. This is
because solidus_i18n uses the :locale param as part of it's path
(ex. /en/products).

For better compatiblity with solidus_i18n, and so that we can merge most
of it's functionality, we should accept either param here (and probably
should prefer :switch_to_locale).