Don't attempt to redirect back in LocaleController
Previously, we attempted to redirect back to the previous URL, and
failing that, the store's root URL.

Redirecting back sounds like a nice feature, but if we are using the
:locale filter from the routing-filter gem (as solidus_i18n does), this
isn't what we want, because the previous URL won't be prefixed by the
current locale.

The existing behaviour of solidus_i18n is already to always redirect to
the root, so we should adopt the same behaviour in solidus itself so
that that override can be removed from solidus_i18n.