Make partially shipped shipment able to be ready
Previously, all inventory units in a shipment needed to be on_hand or
cancelled in order for the shipment to be considered shippable.

This worked fine for most users, but it caused issues for users who were
partially shipping shipments manually using cartons.

This commit allows shipments with some shipped inventory units to still
be considered "ready".