Add initial Spree migration article
This article walks the reader through a simple migrations from Spree
Spree 2.4 to Solidus 1.4. Then, it encourages readers to incrementally
upgrade to newer versions of Solidus.

Both Spree 2.4 and Solidus 1.4 are pretty old now. It would be great to
have an additional article that details a (more complicated) migration
from Spree 3.x to Solidus 2.x. This would necessarily need to be
simplified.