Remove reference to Spree::Shipment.state_machine
There are a few reasons for removing this reference. First, we
should be explicit about which values we want to test; reaching into the
state machine adds an extra layer of abstraction that reduces the
readability of the test. Second, by calling
`Spree::Shipment.state_machine`, we are actually instantiating a state
machine on the `Spree::Shipment` class, which should only ever happen
within the `Shipment` class itself.