Load email validator from core instead of order
It may be desirable to use the email validator from elsewhere before
order.rb has been loaded.

Additionally, this fixes an issue where EmailValidator would be unloaded
in development mode. Spree::Order is autoloaded, and because
Spree::EmailValidator is inside the same namespace, this confuses rails
into thinking Spree::EmailValidator was defined in this file and can be
reloaded. I believe this to be a bug in rails.