Add predicate class for shipments to Spree::Config
Some stores wish to control the building of shipments. Currently this is
done via 'order.has_checkout_step?("delivery")'. This commit adds
a configuration class to control this in order to avoid using the
presence of a delivery step as a proxy for whether or not shipments
should be built.