Improve select2_search_without_selection
Previously this would start failing if the select2 was more than half
full of selected data. This commit changes the selector so that it
always clicks the search portion of the select2 widget.

I'm hopeful that this will also make it more reliable, as specs using
select2_search seem to have been unreliable for the last little bit (but
I have no testing to back up this hope).