Remove assert_written_to_cache
This is error prone and breaks in Rails 5.2, since the key being logged
is an array.

The key we see from AS::Notifications is now an array and has additional
hashes on the template names (I assume russian doll caching related).
It's too hard to predict where these values come from, so we can't
continue testing this.

We should have most of the benefit and testing of this by just counting
the number of cache writes (which we were doing already).