Extend reimbursement feature spec
Reimbursements only work if there is a user with the email address
'spree@example.com', or if you explicitly pass the creator when creating
a reimbursement.

This commit adds to the testing set-up for this spec by mocking
a signed in admin creating the reimbursement, who can be passed as the
creator.