Remove "Path" activation option from promotions UI
While having a promotion activate when the customer reaches a path is a
great feature idea, this this activation type is not actually wired up
to anything. If a store admin tried to set up a path-activated
promotion, it would not work.

This also does not provide a clear path forward for developers trying to
set up a feature like this.

This feature may be a good candidate for an extension with additional
documentation.