Fix security vulnerabilities in docs site
Hoek v2.16.3 has a security vulnerability. Browser-sync and node-sass
were both dependent on this package. I've updated browser-sync to the
latest version. Regarding node-sass, there is not currently a released
version that fixes this vulnerability. They plan on deploying a fix in
version 5, which is supposedly coming in a few weeks. I've gone ahead
and upgraded us to the v5 branch of node-sass for the time being; this
removes our dependency on Hoek, while still allowing the project to
build.