Remove x-overflow rule from admin nav container
This CSS rule is not needed.

There is JS to handle overflowing tabs inside the Solidus Project that
currently resides in:

https://github.com/solidusio/solidus/blob/master/backend/app/assets/javascripts/spree/backend/components/tabs.js

Furthermore this CSS can cause issues for views. When the .content-main
div does not have enough content to be tall enough for the dropdown tab to
fully display a scroll bar is introduced on the right hand side of the page
and looks silly.