Prepare guides for multiple tables of contents
Right now we have published the Solidus Developers Guide. In the future,
there will be additional guides: the Users Guide is currently being
built up in `/guides/source/users/` for example.

This commit removes some of the Developers Guide-specific functionality
of the sidebar navigation anticipating additional TOCs for different
groups of content.