Add helpers to render multiple tables of contents
This commit adds helper methods that will make rendering different
tables of contents on different parts of the site easy.

This also updates the discover_title helper to account for multiple
tables of contents.

\## Dealing with multiple tables of contents

For example, if I visit the following URL:

`localhost:4567/developers/orders/overview.html`

I should see the table of contents for the "Solidus Developers Guide".
And if I visit this URL:

`localhost:4567/users/promotions/overview.html`

I should see the table of contents for the "Solidus Users Guide".

Here's a description of the helpers's functionality:

\## table_of_contents

The table_of_contents helper allows you to determine which table of
contents should be rendered. If the user is viewing a URL with the
subdirectory `/developers/`, then the current table of contents will be
at the `/data/nav/developers.yml` YAML file.

\## current_guide

the current_guide helper finds which guide the user is viewing. If they
are a URL with the subdirectory `/developers/`, then the current guide
is the "Solidus Developers Guide".

The current_guide returns the YAML at `/data/nav/global.yml` which
allows you to get the `current_guide.title` and `current_guide.href` for
the current guide.