Render multiple tables of contents
This commit enables multiple table of contents rendering for the Solidus
Guides.

Each table of contents can have as many "submenus" as you want. Here is
the YAML structure for a short submenu:

```
-
  - title: "Section name"
    dropdown:
      - title: "Article name"
        href: "/relative/link/to/article.html"
      - title: "Another article name"
        href: "/relative/link/to/another-article.html"
```

Things you should note about the structure:

- Each submenu is represented by a root `-` node that is unnamed.
- Each submenu can have multiple sections and multiple articles.
- Articles need to be included in a `dropdown:` at this point in time.

\## Global and Meta tables of contents

The `/data/nav/global.yml` and `/data/nav/meta.yml` do not follow this
data structure.  These are special menus that will be used throughout
the site and should not be rendered only when the user is on specific
pages.