Style links to table of contents indexes
Each table of contents provides a link to the `current_guide.href`. In
the case of the "Solidus Developers Guide" table of contents, that links
to `/developers/index.html`.

These links should be styled to look like the <h1> of the table of
contents block. This is similar to how the title of a blog is usually
the color of a heading or the site's logo, not the color of a link.