Improve UX of "Stores" admin settings UI
This commit improves the user experience of the "Stores" page in the
admin interface in the following ways:

- It clarifies what the "Code" field does.
- It renames the the code field "Slug" in the UI only. This is less
  ambiguous for users who are creating a new store and don't know what
  to put in this field.
- It ensures that all of the fields on the page have the same
  capitalization style.

Note: The UX is still pretty bad. We do not want admins to think making
changes to stores is trivial. We should rethink this page and make more
major changes to it.

Mark "Name" as required for new shipping methods
-----------------------------------------------------------------------

When a store admin creates a new shipping method, the "Name" field
should be marked as required.

This commit also cleans up some extra spacing in the new shipping method
form and fixes the data hook for `admin_name`, which was mis-named to
say `internal_name`.

Update locales strings
-----------------------------------------------------------------------

This commit removes unused locales strings.