Set dummy app forgery protection to true
This is how new (Rails 5.2+) apps are generated by default starting
from rails/rails@ec4a836 commit.

This change makes api specs fail since json POST requests are now
under forgery protection but we have a token based authentication
in place that should replace this check. This will be handled in
a next commit.

We are applying this changes only when rails version is >= 5.2