Fix specs that fail under Rails 5.2.1
I'm not sure why but previously the association was returning an
instance of ActiveRecord::Associations::CollectionProxy that was
considered not changed while reloading. Despite that `shipments`
call is still returning the same object, they now have a different
`object_id` and this could be the root of the problem.

I'm not sure if this issue raised up with Rails 5.2.1 or by using the
ransack master branch on github.