Add a form to create promotion code
Let users explicitly create a new code on an existing promotion,
without checking if it's a "single" or "multiple" promotion type
because it's already possible do switch from a "single" to "multiple"
using the `Spree::PromotionCodeBatch` creation process.

This let the user creates a new code without using the "batch" method
so that it can explicitly choose a new code without using random
suffixes.