Add a Master SKU field to the products form
I have come across a use case in which the product's sku attribute needs to be
editable by the admin even in the event that the product has variants. This
commit adds a user interface element for a Master SKU which updates the
Spree::Product's sku attribute. It is visible regardless of the number of
variants, which satisfies my use case.